Deleting blobs based on dates.

+++ DeleteBlobsTimerTriggerFunc.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
+using System.Configuration;
 using System.Text;
 
 namespace AzFunations
@@ -18,13 +19,21 @@ namespace AzFunations
         {
             log.LogInformation($"C# Timer trigger function executed at: {DateTime.Now}");
 
+
+           var estimatedDays = Environment.GetEnvironmentVariable("BlobEstimatedDays");
+           
+
             BlobContainerClient blobContainerClient =
                 new BlobContainerClient("Connectionstring", "blobContainerName");
 
             var blobs = blobContainerClient.GetBlobs();
             foreach (BlobItem blobItem in blobs)
             {
-                blobContainerClient.DeleteBlobIfExistsAsync(blobItem.Name);
+                TimeSpan difference =  DateTime.Now - Convert.ToDateTime(blobItem.Properties.LastModified);
+                if (difference.Days >= Convert.ToInt32(estimatedDays))
+                {
+                    blobContainerClient.DeleteBlobIfExistsAsync(blobItem.Name);
+                }
                 log.LogInformation($"Blob Name {blobItem.Name} is deleted successfully.");
             }
         }
